Added intInput.h to input directory in utils
Added intInput.h to input directory in utils which contains a function "intInput".
This function's parameters are:

* prompt (string)
* dest (reference of an integer variable)

It will first output the prompt given. Next it will get an input from
the user using getline.
Because of the while loop, it will check if it's a valid integer.
If it's an invalid integer, it will keep getting input
from the user until the response given is a valid integer.
And in each iteration of the loop, it sets the "dest" variable
to the integer given if it's valid.