Added charForms.h to data directory
Added charForms.h to data directory which contains a map "mapCharForms".

This map's key is a char, which acts as a password character.

And this map's value is a map that takes a string and a vector of
characters as key-value pairs, which acts as a map which keys are
categories of password characters (number, uppercase, lower & special),
and values are the different forms of the password character that are in
that category of password characters.