Added functionality for alter method in main.cpp
There are three variables that will be used:

* password (string)
* method (string)
* passIncludeChar (struct includeChar)

It will first ask the user for a method to generate the password and
store that response to the variable "method".

Next it will check if the user entered "alter".
If the user did, it then will ask the user for the
password to alter, and generate a new password using the "methodAlter" function
with the password from the user.
And the result will be stored to the variable "password".

Finally, it will output the generated password.