Added functionality for alter method in main.cpp
There are three variables that will be used:

* password (string)
* method (string)
* passIncludeChar (struct includeChar)

First it will ask the user for a method to generate the password and
store that response to the variable "method".

Next it will check if the user entered "random".
If the user did, it will then ask the user for the length
of the random password, and generate a new password using the
"methodAlter" function with the specified length from the user.
And the result will be stored to the variable "password".

Finally, it will output the generated password.